feat: page for recapitulatif with dynamic segment
