+++ packages/app/src/services/apiClient/siren.ts
@@ -1,6 +1,7 @@
 import moize from "moize";
 
 import { fetcher } from "./fetcher";
+import type { Entreprise } from "@common/models/representation-equilibree";
 
 export type EntrepriseType = {
   adresse?: string;
@@ -14,7 +15,7 @@ export type EntrepriseType = {
   siren?: string;
 };
 
-export const formatAdresse = (entreprise: EntrepriseType) => {
+export const formatAdresse = (entreprise: Entreprise | EntrepriseType) => {
   return [entreprise.adresse, entreprise.code_postal, entreprise.commune];
 };
